Added manager which processes the etcd jobs.
This patch adds a manager to node agent that moniters the
etcd job queue and tries to pick up new jobs. Once it picks
up a job, it tries to execute that job and it updates the
result to the job in the queue.

This patch also adds a sample job executer for command
execution which can be used to test this framework.

Issue #15
Issue #2

Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>