Add service detection feature to node agent
This patch adds service detection capability to node-agent.
Node-agent will have a list of predefined services, which
it will look out for on the node in which it is running. It
will update etcd with the list of services available and or
running on the node. Also based on if the service is running
it will update the TAGS field in the node_context.

tendrl-bug-id: Tendrl/node_agent#136
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>