Fix brick_disk_stats collectd plugin
Brick Disk stats collectd plugin sums all partitions for a partitioned disk stats.
A brick could also be created out of 1 or more but not all partitions of a disk
and summing all partitions stats to get disk stats for a brick might not be very
appropriate. Use the partitions synced by gluster-integration instead.

tendrl-bug-id: Tendrl/node-agent#628
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>