Claim provisioner tag only by managed nodes if cluster is managed
In case cluster is un-managed, any node ideally should be able to
claim the provisioner tag. But if cluster is managed and there are
new nodes for expansion in the cluster, only the managed nodes should
be able to claim the provisioner as new nodes if claim the same there
would be issue while expansion of the cluster.

tendrl-bug-id: Tendrl/node-agent#840
bugzilla: 1599634
Signed-off-by: Shubhendu <shtripat@redhat.com>