Fix service failure in vmware
Currently tendrl using combination of "vendorname + devicetype + serial_id" to
uniquely identify a device. But In a vmware environment some of the devices are
having identical serial_ids and some are very different and having lot of
unicode characters which leads to ommission of identical data which
further leads to missing of some device details to etcd.

The sample serial ids of the devices are as follows in the vmware instances
Serial ID: "`\xbc~\x96\xff\xff\xff\xff\x02"
Serial ID: "./../../../../../bus/scsi"
Serial ID: "./../../../../../bus/scsi"

So there are two errors observed in the vmware environent. One is related to
lengthly unicode characters and missing of data. This is fixed by using
hardware_id of the device in place of serial id.

This patch uses parent-id + unique-id combination used in
place of serial-id of a device to represent as a unique device id
in the VMware environment.

Signed-off-by: Timothy Asir Jeyasingh <tjeyasin@redhat.com>